+++ payment-service/src/main/java/com/flight_booking/payment_service/infrastructure/configuration/kafka/ProducerApplicationKafkaConfig.java
@@ -1,4 +1,4 @@
-package com.flight_booking.booking_service.infrastructure.config;
+package com.flight_booking.payment_service.infrastructure.configuration.kafka;
 
 import com.flight_booking.common.presentation.global.ApiResponse;
 import java.util.HashMap;
